Crate atomic_parser

A crate to parse files using atoms, or to create atoms.

What is an atom?

Atoms are a way for a file to store data. A file like this can be represented as a (real-life) folder with pieces of paper. Each paper has a title (a name) and contents (a payload). Say there are these papers in the folder:

John Doe
foo bar
Ryan Smith

You can find John Doe’s paper in the folder, no matter the order or what other papers there are. In files, atoms are made up of 3 parts:

size
name
payload

The name is the paper title, payload is what is written on the paper. Size is the size of the atom, so the parser knows where each atom ends and a new atom begins.

An example of a file format that works like this is MP4. Note that this parser is not recommended for parsing mp4.

Example

Parsing atoms

Make sure you have the read feature enabled, though it is enabled by default.

Minimal example:

use std::fs::File;
use std::io::Read;
use atomic_parser::Atom;

let mut file = File::open("input.hex").expect("failed to open file.");
let mut value = vec![];
file.read_to_end(&mut value).expect("failed to read file.");
println!("{:?}", Atom::parse(&value, &[]).expect("Atoms are incorrect"));

Or, if we want to ignore skip atoms:

use std::fs::File;
use std::io::Read;
use atomic_parser::Atom;

let mut file = File::open("input.hex").expect("failed to open file.");
let mut value = vec![];
file.read_to_end(&mut value).expect("failed to read file.");
println!("{:?}", Atom::parse(&value, &[[b's', b'k', b'i', b'p']]).expect("Atoms are incorrect"));

We can also ignore multiple atoms, e.g. a skip and a meta atom:

use std::fs::File;
use std::io::Read;
use atomic_parser::Atom;

let mut file = File::open("input.hex").expect("failed to open file.");
let mut value = vec![];
file.read_to_end(&mut value).expect("failed to read file.");
println!("{:?}", Atom::parse(&value, &[[b's', b'k', b'i', b'p'], [b'm', b'e', b't', b'a']]).expect("Atoms are incorrect"));

Building atoms

Make sure you have the write feature enable. Run:

cargo add atomic_parser -F write

Minimal example:

use atomic_parser::Atom;
use atomic_parser::atoms_into_bytes;

let input = vec![
    Atom {
        name: [b'a', b't', b'0', b'1'],
        payload: vec![
            0x01, 0x02, 0x03, 0x04,
        ],
    },
    Atom {
        name: [b'a', b't', b'0', b'2'],
        payload: vec![
           0x01, 0x02, 0x03, 0x04,
        ],
    },
    Atom {
        name: [b'a', b't', b'0', b'3'],
        payload: vec![
            0x01, 0x02, 0x03, 0x04,
        ],
    },
    Atom {
        name: [b'a', b't', b'0', b'4'],
        payload: vec![
            0x01, 0x02, 0x03, 0x04,
        ],
    },
    Atom {
        name: [b'a', b't', b'0', b'5'],
        payload: vec![
            0x01, 0x02, 0x03, 0x04,
        ],
    }
];

let bytes = atoms_into_bytes(input);
println!("{:#?}", bytes);

Structs

Atom

A parsed atom.

AtomError

An error while parsing the atoms.

Enums

AtomErrorKind

Functions

atoms_into_bytes

Turns a vector of atoms into a vector of bytes.