Crate atomic_parser

A crate to parse files using atoms.

Atoms are a way for a file to store data. A file like this can be represented as a (real-life) folder with pieces of paper. Each paper has a title (a name) and contents (a payload). Say there are these papers in the folder:

John Doe
foo bar
Ryan Smith

You can find John Doe’s paper in the folder, no matter the order or what other papers there are. An example of a file format that works like this is MP4. Note that this parser is not recommended for parsing mp4.

Example

use std::fs::File;
use std::io::Read;
use atomic_parser::Atom;

let mut file = File::open("input.hex")?;
let mut value = vec![];
file.read_to_end(&mut value)?;
println!("{:?}", Atom::parse(&value, &[])?);
Ok(())

Structs

Atom

A parsed atom.

AtomError

An error while parsing the atoms.

Enums

AtomErrorKind